refactor: share the pipeline primitives, split the decoder's Vulkan half
Both directions now run a pipelined submit loop, which made the genuinely
common parts visible. Rather than a second copy of each, they move to `video`
alongside `SlotSync`:

- `TimelineChain` replaces the hand-rolled semaphore/next/last triple in three
  places (encode submissions, decode submissions, reorder copies). It separates
  reserving a signal value from committing it, because advancing the chain for a
  submit that failed would leave every later submission waiting on a value
  nothing signals. The encoder had that ordering right and the decoder now
  cannot get it wrong.
- `create_command_pool`, `allocate_command_buffers` and `create_fence` replace
  four hand-rolled copies of the same Vulkan boilerplate across both pipelines
  and both directions' setup.

`decoder/codec.rs` had grown to 1815 lines, well past what AGENTS.md asks for,
and mixed three concerns. Split along its seams, and renamed to `common` so
`codec` is free for the codec trait that mirrors `encoder::codec`:

- `decoder/common.rs` (563): the video session and the per-decoder state.
- `decoder/frames.rs` (550): frame ownership, pool images, DPB slot pins, and
  display-order reordering.
- `decoder/transfer.rs` (716): readback and copies, which share the property of
  running on the transfer queue rather than the decode queue.

No behaviour change, and verified as such: byte-identical output on both GPUs,
display and decode order, validation layers enabled and silent. Intel Arc A310
(ANV) and Radeon RX 9060 XT (RADV) agree with each other and with
`ffmpeg -pix_fmt nv12` on all four streams.
